backend/requirements_audio.txt
# Audio dependencies for voice conversation
deepgram-sdk>=3.0.0
sounddevice>=0.4.0
pygame>=2.0.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
